+++ tarantool/src/net_box/protocol.rs
@@ -1,5 +1,4 @@
 use std::cmp::min;
-use std::fmt::{Display, Formatter};
 use std::io::{self, Cursor, Read, Seek, Write};
 use std::os::raw::c_char;
 use std::str::from_utf8;
@@ -8,6 +7,7 @@ use crate::error::Error;
 use crate::index::IteratorType;
 use crate::msgpack;
 use crate::network::protocol::codec::IProtoType;
+use crate::network::protocol::ResponseError;
 use crate::tuple::{ToTupleBuffer, Tuple};
 
 const REQUEST_TYPE: u8 = 0x00;
@@ -527,15 +527,4 @@ pub fn value_slice(cursor: &mut Cursor<impl AsRef<[u8]>>) -> crate::Result<&[u8]
     let start = cursor.position() as usize;
     msgpack::skip_value(cursor)?;
     Ok(&cursor.get_ref().as_ref()[start..(cursor.position() as usize)])
-}
-
-#[derive(Debug, thiserror::Error)]
-pub struct ResponseError {
-    pub(crate) message: String,
-}
-
-impl Display for ResponseError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{}", self.message)
-    }
 }